tests/wpt/web-platform-tests/css/css-color-adjust/parsing/color-scheme-invalid.html

<!DOCTYPE html>
<title>CSS Color Adjust: parsing color-scheme with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-color-adjust/#color-scheme-prop">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<script>
  test_invalid_value("color-scheme", "none");
  test_invalid_value("color-scheme", "only");
  test_invalid_value("color-scheme", "only only");
  test_invalid_value("color-scheme", "only normal");
  test_invalid_value("color-scheme", "only light dark");
  test_invalid_value("color-scheme", "only light light");
  test_invalid_value("color-scheme", "only light purple");
  test_invalid_value("color-scheme", "normal dark");
  test_invalid_value("color-scheme", "normal purple");
  test_invalid_value("color-scheme", "light normal");
  test_invalid_value("color-scheme", "light none");
  test_invalid_value("color-scheme", "light, dark");
  test_invalid_value("color-scheme", "light inherit");
  test_invalid_value("color-scheme", "light initial");
  test_invalid_value("color-scheme", "light unset");
  test_invalid_value("color-scheme", "light revert");
  test_invalid_value("color-scheme", "light default");
</script>
